balance_ai/pyproject.toml
[project]
name = "balance-ai"
version = "0.1.0"
description = "AI optimizer for Bastion's Last Stand game balance"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = []
